[Potentially breaking commit] - Refactor of the output of the forward method of the emotion losses (before it was just the forward pass of the emotion net, now it's a computation of the loss as a whole. Potentially breaking some old stuff, but I don't think so. - adding a configurable keyword to the renderer (for instance "image" vs "video" - weird thing in mouth crop function. The grid size had mismatched number of dimensions but addition worked anyway but caused a weirdness. Not sure why it's coming up now and not before. Added an if to fix the dimensions if mismatched but this could be cleaner.
The forward steps of new face reconstruction now runs.

+++ gdl/datasets/VideoDatasetBase.py
@@ -699,7 +699,7 @@ class VideoDatasetBase(AbstractVideoDataset):
                 landmark_validity = np.concatenate([landmark_validity, np.zeros((sequence_length - landmark_validity.shape[0], 1), 
                     dtype=landmark_validity.dtype)], axis=0)
 
-            landmark_dict[landmark_type] = landmarks
+            landmark_dict[landmark_type] = landmarks.astype(np.float32)
             landmark_validity_dict[landmark_type] = landmark_validity
 
 
@@ -1218,7 +1218,7 @@ class VideoDatasetBase(AbstractVideoDataset):
 
         # split the augmented landmarks back to their original types
         for i, lmk_type in enumerate(sample["landmarks"].keys()):
-            sample["landmarks"][lmk_type] = landmarks_to_augment_aug[:, np.sum(lmk_counts[:i]):np.sum(lmk_counts[:i+1]), :]
+            sample["landmarks"][lmk_type] = landmarks_to_augment_aug[:, np.sum(lmk_counts[:i]):np.sum(lmk_counts[:i+1]), :].astype(np.float32)
         return sample
 
     def _occlude_sequence(self, index, images, segmentation, mediapipe_landmarks, mediapipe_landmarks_valid, region):
